move preprocess into tsr.utils
